requirements.txt
aiogram==2.10.1
aiohttp==3.7.2
alembic==1.4.3
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
autoflake==1.4
Babel==2.8.0
beautifulsoup4==4.9.3
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
flake8==3.8.4
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
invoke==1.4.1
isort==5.6.4
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.0.0
mypy==0.790
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-mock==3.3.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
regex==2020.10.28
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.2
zipp==3.4.0
